Hard assertion that we get the correct return value (not happening on node; only client 1 runs, while client 2 is never woken up)

+++ tau/flux/types.hh
@@ -31,7 +31,7 @@ using tau::util::operator<<;
 
 typedef uN      λi;  // λ identifier
 typedef f64     λp;  // λ priority (lower = more important)
-typedef int     Λr;  // type of value returned to Λ
+typedef i64     Λr;  // type of value returned to Λ
 typedef F<Λr()> λf;
 
 typedef uN      ψi;  // Γ-local port identifier
